==> scripts_manuscript/2_Normalized_to_DE_DESEQ_SVA.Rmd <==
---
title: "Normalized_to_DE with SUV batch corrected data"
author: "Caroline_Duncombe"
date: "2024-11-13"
output: html_document
---

PART II: DSP Analysis on GeoMX data 

Script will take the normalized/batch corrected data and perform DE analysis with DESeq. 

############# A PRIOR POINTS #############

Goal of this script is to create a final usable DE file for down stream analysis. 

Thoughts on the model. I have reviewed other models and have figured out several point.
* Zonation does not alter the measured interferon response around the schizont. Only one DE gene was identified as modified by zone. Will only include as interaction term because of previous precedent in the literature. 
* There is no measurable difference around schizont or bystander region. 
  ****** GOING TO REMOVE BYSTANDER FROM MODEL. 
* ROI size has an impact. Running this 200um ROI to keep consistent. 
* I will have two models. 
    * Interaction_model - Group*Type + Group:Liver_zone + Liver_zone (Main model)
    * Contrast_model -    Group_Type + Liver_zone (Secondary model)
        ***Contrast model is used to makes the Group & Type a unique variable to pull out exact contrasts for Gene Enrichment.
    * Interaction model with bystander - Group*Type + Group:Liver_zone + Liver_zone.
        *** Model is simply to create the Volcano plot for schizont versus bystander figure. 
        *** GW -  would love your input if there is a better way to do this. Running to prove why I chose the first model. 

##################################################################

### One Question for GW:
  Do you use the CCA corrected values for DEseq2? Here I will conduct without the CCA corrected values. 

Load in and normalize data
```{r load the libraies }

library(here)
library(GeomxTools)
library(CellMembrane)
library(DESeq2)
library(Seurat)
library(dplyr)
#load or download the 'usethis' package.
pacman::p_load("usethis")
pacman::p_load(BiocManager, devtools)

rm(list = ls())

seuratObj.tmp <- readRDS("clean_data_manuscript/seurat_object/QC_normalized_data_200um_SVA.rds")

```

########################################################################################################################################

# MODEL 1: INTERACTION MODEL

########################################################################################################################################

RUN & SAVE MODEL 1: INTERACTION TERM


######## QUESTION FOR GW!!!!! ##############

Depending on contrast method I use with "apeglm" or "ashr" in the DESeq run. What is the correct answer here? Is it okay if they are different when I am running the two different models?

#############################################


```{r Group*Type + Group:Liver_zone + Liver_zone}

# SUBSET DATA TO NOT INCLUDE THE BYSTANDER GROUP!!!!

seuratObj.noby <- subset(seuratObj.tmp, subset = Type != "bystander")

# need matrix and metadata:
cts <- Seurat::GetAssayData(seuratObj.noby, assay = "SVA_batchCorrected", layer = "counts")
coldata <- seuratObj.noby@meta.data

#Intercept one - always compared against female
dds.noby <- DESeqDataSetFromMatrix(countData = round(as.matrix(cts),0),
                             colData = coldata,
                             design= ~ Group*Type + Group:Liver_zone + Liver_zone)

dds.noby$Liver_zone <- relevel(dds.noby$Liver_zone, ref = "PP")
dds.noby$Group <- relevel(dds.noby$Group, ref = "M")

dds.noby <- DESeq(dds.noby, minReplicatesForReplace=Inf)

resultsNames(dds.noby) # lists of comparsions:

```


# RUNS FUNCTIONS

These functions will create a DE list for each term, compile into a single df, and then export and save the volcano plots for each. 

######## QUESTION FOR GW!!!!! ##############

Can you review these functions and make sure they look correct!

#############################################


```{r Function - Create interative code to create complete DE list with savable R lists}
#' @title PerformDifferentialExpressionForAllContrasts
#'
#' @description This function performs differential expression analysis using DESeq2 for all contrasts in a pseudobulked Seurat object.
#' @param dds The DESeq2 dataset object (dds) generated by DESeq2.
#' @param logFC_threshold The minimum log2 fold change to report.
#' @param FDR_threshold The maximum false discovery rate (adjusted p-value) to report.
#' @param test.use Can be either "Wald" or "LRT". "Wald" runs DESeq2's Wald test, while "LRT" runs the likelihood ratio test.
#' @param showPlots Boolean determining if the volcano plots and p-value distributions should be shown.
#' @return A named list where each element corresponds to a contrast, containing the differential expression results, a volcano ggplot object, and a p-value distribution ggplot object.
#' @export

PerformDifferentialExpressionForAllContrasts <- function(dds, logFC_threshold = 1, FDR_threshold = 0.05, test.use = "Wald", showPlots = TRUE) {
  # Check if the test type is valid
  if (!test.use %in% c("Wald", "LRT")) {
    stop("Please supply a valid test.use argument: 'Wald' or 'LRT'.")
  }

  # Get all contrast names
  contrast_names <- resultsNames(dds)
  
  # Filter out contrasts starting with "Intercept"
  contrast_names <- contrast_names[!grepl("^Intercept", contrast_names)]
  
  # Debug: Print contrast names
  print("Contrast Names:")
  print(contrast_names)

  # Initialize an empty list to store results for each contrast
  results_list <- list()

  # Loop over each contrast name and perform differential expression analysis
  for (contrast_name in contrast_names) {
    # Perform differential expression analysis for the given contrast
    if (test.use == "Wald") {
      fit_test <- DESeq2::results(dds, name = contrast_name, cooksCutoff = FALSE, independentFiltering=FALSE)
    } else if (test.use == "LRT") {
      fit_test <- DESeq2::results(dds, contrast = contrast_name, test = "LRT", cooksCutoff = FALSE,independentFiltering=FALSE)
    }

    # Check if fit_test has results
    if (is.null(fit_test)) {
      warning(paste("No results for contrast:", contrast_name))
      next
    }

    # Shrink log2 fold changes (optional, recommended for visualization)
    fit_test <- DESeq2::lfcShrink(dds, coef = contrast_name, type = "apeglm")

    # Convert results to a data frame and add a 'gene' column
    differential_expression <- as.data.frame(fit_test)
    differential_expression$gene <- rownames(differential_expression)

    # Check if differential_expression is empty
    if (nrow(differential_expression) == 0) {
      warning(paste("No differential expression data for contrast:", contrast_name))
      next
    }

    # Create volcano plot
    volcano <- ggplot(differential_expression, aes(x = log2FoldChange, y = -log10(padj), label = gene)) +
      geom_point(data = differential_expression[differential_expression$log2FoldChange >= logFC_threshold &
                                                differential_expression$padj <= FDR_threshold,], color = "orange") +
      geom_point(data = differential_expression[differential_expression$log2FoldChange <= -logFC_threshold &
                                                differential_expression$padj <= FDR_threshold,], color = "cadetblue2") +
      ggrepel::geom_text_repel(data = differential_expression[
        (differential_expression$log2FoldChange >= logFC_threshold & differential_expression$padj < FDR_threshold) |
          (differential_expression$log2FoldChange <= -logFC_threshold & differential_expression$padj < FDR_threshold),],
        max.overlaps = 10) +
      geom_point(data = differential_expression[(differential_expression$log2FoldChange > -logFC_threshold &
                                                 differential_expression$log2FoldChange < logFC_threshold) |
                                                 differential_expression$padj > FDR_threshold,], color = "black") +
      geom_hline(yintercept = -log10(FDR_threshold), color = "red", linetype = "dashed") +
      geom_vline(xintercept = logFC_threshold, color = "red", linetype = "dashed") +
      geom_vline(xintercept = -logFC_threshold, color = "red", linetype = "dashed") +
      egg::theme_article() +
    labs(x = "Log2 Fold Change", y = "-Log10 Adjusted P-value")

    # Create p-value distribution plot
    pvalue_dist <- ggplot(differential_expression, aes(x = pvalue)) + 
      geom_histogram() + 
      ggtitle(paste("P-Value Distribution for", contrast_name)) + 
      egg::theme_article()

    # Display plots if requested
    if (showPlots) {
      print(pvalue_dist)
      print(volcano)
    }

    # Store results in the list with the contrast name as the key
    results_list[[contrast_name]] <- list(
      differential_expression = differential_expression,
      volcano = volcano,
      pvalue_dist = pvalue_dist
    )
    
  }

  # Return the list of results
    print(paste("Stored results for contrast:", contrast_name))
    print(results_list)
    return(results_list)
}
```

```{r Function - extract DE list from previous function output}

#' @title CombineDEResults
#'
#' @description This function processes differential expression results from a list by converting rownames to a column, adding a contrast name, and combining all dataframes into a single dataframe.
#' @param de_list A list of differential expression results where each element contains a dataframe with rownames that need to be converted to a column.
#' @return A single dataframe with rownames removed, converted to a column, and a new column for the contrast name.
#' @export
CombineDEResults <- function(de_list) {
  # Initialize an empty list to store dataframes for each contrast
  df_list <- list()
  
  # Loop over each contrast in the input list
  for (contrast_name in names(de_list)) {
    # Extract the differential expression dataframe for the current contrast
    de_df <- de_list[[contrast_name]]$differential_expression
    
    # Convert rownames to a column and add contrast name
    de_df <- as.data.frame(de_df)
    de_df$gene <- rownames(de_df)
    de_df$contrast <- contrast_name
    rownames(de_df) <- NULL
    
    # Append the dataframe to the list
    df_list[[contrast_name]] <- de_df
  }
  
  # Combine all dataframes in the list into a single dataframe
  final_df <- bind_rows(df_list)
  
  return(final_df)
}

```

```{r Function - to save volcano plot images}

#' @title SaveVolcanoPlots
#'
#' @description This function saves volcano plots from a list of differential expression results into a specified directory.
#' @param df_list A named list where each element contains a volcano plot (ggplot object) and has a name corresponding to the contrast.
#' @param output_dir The directory where the images should be saved.
#' @return NULL
#' @export

SaveVolcanoPlots <- function(df_list, output_dir = "figs/DE_volcanoplots/DE.noby.50") {
  # Create the directory if it does not exist
  if (!dir.exists(output_dir)) {
    dir.create(output_dir, recursive = TRUE)
  }
  
  # Loop over each contrast in df_list
  for (contrast_name in names(df_list)) {
    # Get the volcano plot
    p <- df_list[[contrast_name]]$volcano
    
    # Define the file path
    file_path <- file.path(output_dir, paste0(contrast_name, "_SUV.pdf"))
    
    # Save the plot
    ggsave(filename = file_path, plot = p, width = 4, height = 2.7)
  }
}

```

```{r Run this work flow and save the outputs}

DE.noby <- PerformDifferentialExpressionForAllContrasts(dds.noby)

DE.noby.combined <- CombineDEResults(DE.noby)

write.csv(DE.noby.combined, "clean_data_manuscript/DE/DE_nobystander_all_interaction_model.csv", row.names = FALSE)

# Call the function with DE.noby.50
# Example usage:
SaveVolcanoPlots(df_list = DE.noby, output_dir = "figs_manuscript/Volcano_plots/DE.noby_interaction_model")

```




########################################################################################################################################

# MODEL 2: CONTRAST MODEL

########################################################################################################################################

RUN & SAVE MODEL 2: Contrast Term


```{r Group_Type + Liver_zonel}

# Filter out the bystander again. 
seuratObj.noby <- subset(seuratObj.tmp, subset = Type != "bystander")

# need matrix and metadata:
cts <- Seurat::GetAssayData(seuratObj.noby, assay = "SVA_batchCorrected", layer = "counts")

# Make a variable representing the contrasts of interest. 
coldata <- seuratObj.noby@meta.data %>% mutate(Group_Type = paste0(Group,"_",Type))

#Intercept one - always compared against female
dds.noby.contrast <- DESeqDataSetFromMatrix(countData = round(as.matrix(cts),0),
                             colData = coldata,
                             design= ~ Group_Type + Liver_zone)

dds.noby.contrast$Liver_zone <- relevel(dds.noby.contrast$Liver_zone, ref = "PP")
dds.noby.contrast$Group <- relevel(dds.noby.contrast$Group_Type, ref = "M_mock")

dds.noby.contrast <- DESeq(dds.noby.contrast, minReplicatesForReplace=Inf)

resultsNames(dds.noby.contrast)

```



Run the workflow.
```{r Run this work flow and save the outputs for standard terms}

DE.noby <- PerformDifferentialExpressionForAllContrasts(dds.noby.contrast)

DE.noby.combined <- CombineDEResults(DE.noby)

write.csv(DE.noby.combined, "clean_data_manuscript/DE/DE_nobystander_all_contrast_model.csv", row.names = FALSE)

# Call the function with DE.noby.50
# Example usage:
SaveVolcanoPlots(df_list = DE.noby, output_dir = "figs_manuscript/Volcano_plots/DE.noby_contrast_model")

```

I am interested in pulling different contrast terms. 
 
Here I will select the contrasts of interest for downstream analysis
 
```{r Looking at contrasts of interest}

# Extract individual results with specified contrasts
res_F_schizont_vs_F_mock <- results(dds.noby.contrast, contrast = c("Group_Type", "F_schizont", "F_mock"))
res_M_schizont_vs_M_mock <- results(dds.noby.contrast, contrast = c("Group_Type", "M_schizont", "M_mock"))
res_ORX_schizont_vs_ORX_mock <- results(dds.noby.contrast, contrast = c("Group_Type", "ORX_schizont", "ORX_mock"))
res_M_mock_vs_F_mock <- results(dds.noby.contrast, contrast = c("Group_Type", "M_mock", "F_mock"))
res_M_mock_vs_ORX_mock <- results(dds.noby.contrast, contrast = c("Group_Type", "M_mock", "ORX_mock"))
res_F_mock_vs_ORX_mock <- results(dds.noby.contrast, contrast = c("Group_Type", "F_mock", "ORX_mock"))

```

```{r Loop to save DE and volcano plot}


# Define contrast parameters and shrinkage contrasts
contrasts <- list(
  F_schizont_vs_F_mock = c("Group_Type", "F_schizont", "F_mock"),
  M_schizont_vs_M_mock = c("Group_Type", "M_schizont", "M_mock"),
  ORX_schizont_vs_ORX_mock = c("Group_Type", "ORX_schizont", "ORX_mock"),
  M_mock_vs_F_mock = c("Group_Type", "M_mock", "F_mock"),
  M_mock_vs_ORX_mock = c("Group_Type", "M_mock", "ORX_mock"),
  F_mock_vs_ORX_mock = c("Group_Type", "F_mock", "ORX_mock")
)

contrast_names <- c(
  "F_schizont_vs_F_mock",
  "M_schizont_vs_M_mock",
  "ORX_schizont_vs_ORX_mock",
  "M_mock_vs_F_mock",
  "M_mock_vs_ORX_mock",
  "F_mock_vs_ORX_mock"
)
# Volcano plot parameters
logFC_threshold <- 1
FDR_threshold <- 0.05

# Loop over contrasts to create volcano plots and save data
for (i in seq_along(contrasts)) {
  # Set contrast name for folder and file naming
  contrast <- contrasts[[i]]
  contrast_name <- contrast_names[i]
  
  # Perform log fold change shrinkage for visualization
  fit_test <- DESeq2::lfcShrink(dds.noby.contrast, contrast = contrast, type = "ashr")
  
  # Convert results to a data frame and add a 'gene' column
  differential_expression <- as.data.frame(fit_test)
  differential_expression$gene <- rownames(differential_expression)
  
  # Create a folder for each contrast
  output_dir.volcano <- file.path("figs_manuscript/Volcano_plots/DE.noby_contrast_model/")
  output_dir.DE <- file.path("clean_data/DE/DE.noby_contrast_model/")
  dir.create(output_dir.volcano, showWarnings = FALSE, recursive = TRUE)
  dir.create(output_dir.DE, showWarnings = FALSE, recursive = TRUE)
  
  # Generate the volcano plot
  volcano <- ggplot(differential_expression, aes(x = log2FoldChange, y = -log10(padj), label = gene)) +
    # Highlight upregulated genes
    geom_point(data = differential_expression[differential_expression$log2FoldChange >= logFC_threshold &
                                                differential_expression$padj <= FDR_threshold, ], color = "orange") +
    # Highlight downregulated genes
    geom_point(data = differential_expression[differential_expression$log2FoldChange <= -logFC_threshold &
                                                differential_expression$padj <= FDR_threshold, ], color = "cadetblue2") +
    # Add labels for significant genes using ggrepel
    ggrepel::geom_text_repel(data = differential_expression[
      (differential_expression$log2FoldChange >= logFC_threshold & differential_expression$padj < FDR_threshold) |
      (differential_expression$log2FoldChange <= -logFC_threshold & differential_expression$padj < FDR_threshold), ],
      max.overlaps = 10) +
    # Neutral points (non-significant)
    geom_point(data = differential_expression[
      (differential_expression$log2FoldChange > -logFC_threshold & differential_expression$log2FoldChange < logFC_threshold) |
      differential_expression$padj > FDR_threshold, ], color = "black") +
    # Add threshold lines for FDR and log2 fold change
    geom_hline(yintercept = -log10(FDR_threshold), color = "red", linetype = "dashed") +
    geom_vline(xintercept = logFC_threshold, color = "red", linetype = "dashed") +
    geom_vline(xintercept = -logFC_threshold, color = "red", linetype = "dashed") +
    ggtitle(contrast_name) +
    egg::theme_article() +
    labs(x = "Log2 Fold Change", y = "-Log10 Adjusted P-value")
  
  # Save volcano plot as an image
  volcano_filepath <- file.path(output_dir.volcano, paste0(contrast_name, "_volcano_plot.png"))
  ggsave(volcano_filepath, plot = volcano, width = 4, height = 2.7)
  
  # Save differential expression data as a CSV
  data_filepath <- file.path(output_dir.DE, paste0(contrast_name, "_differential_expression.csv"))
  write.csv(differential_expression, data_filepath, row.names = FALSE)
  
  print(paste("Saved volcano plot and differential expression data for contrast:", contrast_name))
}

```


########################################################################################################################################

# Create Volcano plot for Bystander versus Schizont 

########################################################################################################################################

```{r Group*Type + ROI_size + Group:Liver_zone + Liver_zone}

# SUBSET DATA TO NOT INCLUDE THE BYSTANDER GROUP!!!!
seuratObj.by <- subset(seuratObj.tmp)

# need matrix and metadata:
cts <- Seurat::GetAssayData(seuratObj.by, assay = "SVA_batchCorrected", layer = "counts")
coldata <- seuratObj.by@meta.data

#Intercept one - always compared against female
dds.by <- DESeqDataSetFromMatrix(countData = round(as.matrix(cts),0),
                             colData = coldata,
                             design= ~ Group*Type + Group:Liver_zone + Liver_zone)

dds.by$Liver_zone <- relevel(dds.by$Liver_zone, ref = "PP")
dds.by$Group <- relevel(dds.by$Group, ref = "M")

dds.by <- DESeq(dds.by, minReplicatesForReplace=Inf)

resultsNames(dds.by) # lists of comparsions:

```

```{r Run this work flow and save the outputs.}

DE.by <- PerformDifferentialExpressionForAllContrasts(dds.by)

DE.by.combined <- CombineDEResults(DE.by)

write.csv(DE.by.combined, "clean_data_manuscript/DE/DE_with_bystander_all_interaction_model.csv", row.names = FALSE)

# Call the function with DE.noby.50
# Example usage:
SaveVolcanoPlots(df_list = DE.by, output_dir = "figs_manuscript/Volcano_plots/DE.with_by_interaction_model")

```

##### ALL DONE!!! #######
